requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-bootstrap-datepicker-plus==3.0.5
django-choices==1.7.1
django-countries==6.1.3
django-crispy-forms==1.9.2
django-filter==2.4.0
django-tinymce==3.1.0
djangorestframework==3.12.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
